update: move files
- slit: code in `main.rs` file
- add: sqlite pool
