+++ .gitignore
@@ -15,3 +15,4 @@ Cargo.lock
 
 .vscode
 .env
+*.db
