+++ Cargo.toml
@@ -14,21 +14,20 @@ leptos_router = { version = "0.7.0" }
 leptos_axum = { version = "0.7.0", optional = true }
 #
 wasm-bindgen = { version = "=0.2.100", optional = true }
-console_error_panic_hook = { version = "0.1", optional = true}
+console_error_panic_hook = { version = "0.1", optional = true }
 tokio = { version = "1", features = ["rt-multi-thread"], optional = true }
 #
 chrono = "0.4.39"
 leptos_calendar = "0.1.2"
-clap = { version = "4.5.28", features = ["derive", "env"] }
-sqlx = { version = "0.8.2", features = ["runtime-tokio", "sqlite"], optional = true }
 serde = { version = "1.0.217", features = ["derive"] }
+clap = { version = "4.5.28", features = ["derive", "env"] }
+sqlx = { version = "0.8.2", features = [
+    "sqlite",
+    "runtime-tokio",
+], optional = true }
 
 [features]
-web = [
-    "leptos/hydrate",
-    "dep:console_error_panic_hook",
-    "dep:wasm-bindgen",
-]
+web = ["leptos/hydrate", "dep:console_error_panic_hook", "dep:wasm-bindgen"]
 server = [
     "dep:axum",
     "dep:sqlx",
