+++ env.example
@@ -1,0 +1,2 @@
+MRP_MODE=
+MRP_DATABASE_PATH=
