+++ src/interfaces/mod.rs
@@ -1,0 +1,3 @@
+pub mod cli;
+#[cfg(feature = "server")]
+pub mod web;
