+++ src/lib.rs
@@ -1,14 +1,17 @@
 mod api;
 pub mod app;
 mod calendar;
-mod cli;
+mod interfaces;
 mod pages;
 mod services;
 
 // exposes needed functions to binary
 pub mod shared {
-    pub use crate::cli::*;
+    pub use crate::interfaces::cli::*;
+    #[cfg(feature = "server")]
+    pub use crate::interfaces::web::*;
     pub use crate::services::*;
+    pub use crate::shared::database::*;
 }
 
 pub type AppError = String;
@@ -18,13 +21,13 @@ pub type AppResult<T> = std::result::Result<T, AppError>;
 #[wasm_bindgen::prelude::wasm_bindgen]
 pub fn hydrate() {
     use crate::app::*;
-    use crate::cli::*;
+    use crate::interfaces::cli::*;
 
     console_error_panic_hook::set_once();
-    let AppCli { mode } = AppCli::parse();
+    let AppCli { mode, .. } = AppCli::parse();
 
     match mode {
-        Modes::Private => leptos::mount::hydrate_body(AdminApp),
-        Modes::Public => leptos::mount::hydrate_body(App),
+        Modes::Server => leptos::mount::hydrate_body(AdminApp),
+        Modes::Client => leptos::mount::hydrate_body(App),
     };
 }
